move user command to own cog
