+++ bot/cogs/utilities.py
@@ -2,42 +2,21 @@ from asyncio import sleep
 from pathlib import Path
 import random
 import subprocess
-import datetime
 from typing import Optional
 
-from discord import ActivityType, Colour, Embed, Member, Spotify, Status
+from discord import Colour, Embed
 from discord.ext import tasks
 from discord.ext.commands import Cog, Context, command
 import yaml
 
 from bot import settings
 from bot.bot import Friendo
-from dateutil.relativedelta import relativedelta
 
 with open(Path.cwd() / 'bot' / 'resources' / 'list_of_quotes.yaml', 'r', encoding='utf-8') as f:
     lines = yaml.load(f, Loader=yaml.FullLoader)['lines']
 
 # Define the time period units user can pass
 VALID_PERIODS = "s sec secs second seconds m min mins minute minutes h hour hours".split()
-
-# Dictionaries for the emojis in userinfo embed
-BADGES = {'hypesquad_bravery': '<:bravery:784430728683323392>',
-          'hypesquad_brilliance': '<:brilliance:784430705118806036>',
-          'hypesquad_balance': '<:balance:784430681857720332>',
-          'nitro': '<:nitro:784751384780079134>',
-          'verified_bot_developer': '<:dev:784755774781390849>',
-          'partner': '<:partner:784756644634034176>',
-          'early_supporter': '<:early:784756624509108224>'}
-
-STATUSES = {Status.online: '<:online:782636673653407744>',
-            Status.offline: '<:offline:782636856075616286>',
-            Status.idle: '<:3929_idle:782638272483950623>',
-            Status.dnd: '<:dnd:782637101464027136>'}
-
-ACTIVITIES = {ActivityType.playing: ':video_game: Playing ',
-              ActivityType.listening: ':headphones: Listening to ',
-              ActivityType.streaming: ':desktop: Streaming on ',
-              ActivityType.custom: ''}
 
 
 def convert_time(time: str, period: str) -> Optional[int]:
@@ -72,35 +51,6 @@ class Utilities(Cog):
         self.drink_tasks = {}
         self.reminder_tasks = {}
         self.reminder_limit = 1
-
-    @staticmethod
-    def get_timedelta(a: datetime.datetime, b: datetime.datetime) -> list:
-        """Static method for getting parsed time delta between 2 datetimes."""
-        final = []
-        delta = relativedelta(a, b)
-        years = abs(delta.years)
-        months = abs(delta.months)
-        days = abs(delta.days)
-        hours = abs(delta.hours)
-        minutes = abs(delta.minutes)
-
-        if minutes and not months and not years:
-            final.append(f'{minutes} minutes')
-
-        if hours:
-            final.append(f'{hours} hours')
-
-        if days:
-            final.append(f'{days} days')
-
-        if months:
-            final.append(f'{months} months')
-
-        if years:
-            final.append(f'{years} years')
-        final = final[::-1]
-
-        return final[:3]
 
     async def reminder_wrapper(
             self,
@@ -232,77 +182,6 @@ class Utilities(Cog):
         embed_quote = Embed(title=random.choice(lines), color=Colour.green())
 
         await ctx.send(embed=embed_quote)
-
-    @command(brief="Get the info on the user specified, defaults to the command author",
-             usage=".userinfo [user (optional)]",
-             aliases=('ui', 'user', 'useri', 'uinfo'))
-    async def userinfo(self, ctx: Context, user: Member = None) -> None:
-        """Shows an embed containing basic info on the user."""
-        user = ctx.author if not user else user
-        spotify_emoji = '<:spotify:785113868543852584>'
-        user_badges = []
-        roles = []
-        statuses = []
-        is_bot = "Bot: :x:"
-
-        create_time = ', '.join(self.get_timedelta(datetime.datetime.utcnow(), user.created_at))
-        joined_time = ', '.join(self.get_timedelta(datetime.datetime.utcnow(), user.joined_at))
-
-        flags = user.public_flags
-
-        info_emb = Embed(color=user.color,
-                         title=str(user))
-        info_emb.set_thumbnail(url=user.avatar_url)
-
-        for flag_ in flags.all():
-            if str(flag_.name) in BADGES:
-                user_badges.append(BADGES[str(flag_.name)])
-
-        if user.is_avatar_animated():
-            user_badges.append(BADGES['nitro'])
-
-        for role in user.roles[1:]:
-            roles.append(f"<@&{role.id}>")
-
-        if user.bot:
-            is_bot = "Bot: :white_check_mark:"
-
-        statuses.append(f"{STATUSES[user.mobile_status]} Mobile Client")
-        statuses.append(f"{STATUSES[user.web_status]} Web Client")
-        statuses.append(f"{STATUSES[user.desktop_status]} Desktop Client")
-
-        roles = ', '.join(roles)
-        user_badges = " ".join(user_badges)
-        statuses = '\n'.join(statuses)
-        info_emb.description = user_badges
-        men, id_, nick = user.mention, user.id, user.nick
-
-        base_activity = user.activity
-        if base_activity:
-            activity = ACTIVITIES[base_activity.type] + base_activity.name
-        else:
-            activity = "No activity is being done"
-
-        if isinstance(base_activity, Spotify):
-            activity = "\n".join([f'{spotify_emoji} {activity[12:]}',
-                                  f"**Song**: {base_activity.title}",
-                                  f"**Artist**: {base_activity.artist}",
-                                  f"**Album**: {base_activity.album}"])
-
-        info_emb.add_field(name='User Information',
-                           value=f"Created: {create_time} ago\nProfile: {men}\nID: {id_}\n{is_bot}",
-                           inline=False)
-        info_emb.add_field(name="Guild Profile",
-                           value=f"Joined: {joined_time} ago\nNick: {nick}\nRoles: {roles}",
-                           inline=False)
-        info_emb.add_field(name="Status",
-                           value=statuses,
-                           inline=False)
-        info_emb.add_field(name="Activity",
-                           value=activity,
-                           inline=False)
-
-        await ctx.send(embed=info_emb)
 
 
 def setup(bot: Friendo) -> None:
